+++ src/main/java/de/nordakademie/iaa/controller/RoomController.java
@@ -1,9 +1,10 @@
 package de.nordakademie.iaa.controller;
 
 
+import de.nordakademie.iaa.model.Event;
 import de.nordakademie.iaa.model.Room;
+import de.nordakademie.iaa.service.EventService;
 import de.nordakademie.iaa.service.RoomService;
-import de.nordakademie.iaa.service.exception.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -20,10 +21,12 @@ import static org.springframework.web.bind.annotation.RequestMethod.PUT;
 @RequestMapping("/rooms")
 public class RoomController {
 
+    private EventService eventService;
     private RoomService roomService;
 
     @Autowired
-    public RoomController(RoomService roomService) {
+    public RoomController(EventService eventService, RoomService roomService) {
+        this.eventService = eventService;
         this.roomService = roomService;
     }
 
@@ -49,7 +52,8 @@ public class RoomController {
                 roomService.saveRoom(room);
                 return ResponseEntity.status(HttpStatus.CREATED).build();
             }
-        } catch (Exception ignored) {}
+        } catch (Exception ignored) {
+        }
         return ResponseEntity.badRequest().build();
     }
 
@@ -65,7 +69,8 @@ public class RoomController {
                 roomService.saveRoom(room);
                 return ResponseEntity.ok().build();
             }
-        } catch (Exception ignored) {}
+        } catch (Exception ignored) {
+        }
         return ResponseEntity.badRequest().build();
     }
 
@@ -76,12 +81,20 @@ public class RoomController {
      */
     @RequestMapping(value = "/{id}", method = DELETE)
     public ResponseEntity deleteRoom(@PathVariable Long id) {
-        try {
+        Room room = roomService.loadRoom(id);
+        if (room != null) {
+            List<Event> eventsWithRoom = eventService.findEventsByRoom(room);
+            eventsWithRoom.forEach(event -> {
+                if (event.getRooms().size() == 1) {
+                    eventService.deleteEvent(event.getId());
+                } else {
+                    event.removeRoom(room);
+                }
+            });
             roomService.deleteRoom(id);
             return ResponseEntity.ok(null);
-        } catch (EntityNotFoundException e) {
-            return ResponseEntity.badRequest().build();
         }
+        return ResponseEntity.badRequest().build();
     }
 
 }
